Fork sync (#2)
* Add more HTTP/2 tests for RFC7540 4.1 and 4.3

* Add a guide chapter linking to all specs

* Add the early_error cowboy_stream callback

This callback is called when an error occurs before the request
(including headers, excluding body) was fully received. The
init/3 callback will not be called. The callback receives the
partial Req object (possibly empty), the reason for the error
and the response command that the server will send. It allows
you to be aware of the error and possibly modify the response
before it is sent.

* Initial cowboy_stream manual